nn/requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
pyyaml
torchview
graphviz
